// interfaces/image.ts
export interface ImageResponse {
  total: number
  totalHits: number
  hits: PixabayImage[]
}

export interface PixabayImage {
  id: number
  pageURL: string
  type: "photo"
  tags: string
  previewURL: string
  previewWidth: number
  previewHeight: number
  webformatURL: string
  webformatWidth: number
  webformatHeight: number
  largeImageURL: string
  fullHDURL: string
  imageURL: string
  imageWidth: number
  imageHeight: number
  imageSize: number
  views: number
  downloads: number
  likes: number
  comments: number
  user_id: number
  user: string
  userImageURL: string
}